requirements.txt
openai==1.30.3
spotipy==2.22.1
Pillow==10.3.0